move leia displaymanager calls to gameview
